+++ sqlparser/lib/src/engine/sql_engine.dart
@@ -111,6 +111,12 @@ class SqlEngine {
     return scope;
   }
 
+  Scanner _scan(String source) {
+    final scanner =
+        Scanner(source, scanDriftTokens: options.useDriftExtensions);
+    return scanner..scanTokens();
+  }
+
   /// Tokenizes the [source] into a list list [Token]s. Each [Token] contains
   /// information about where it appears in the [source] and a [TokenType].
   ///
@@ -119,30 +125,42 @@ class SqlEngine {
   /// you need to filter them. When using the methods in this class, this will
   /// be taken care of automatically.
   List<Token> tokenize(String source) {
-    final scanner =
-        Scanner(source, scanDriftTokens: options.useDriftExtensions);
-    final tokens = scanner.scanTokens();
+    final scanner = _scan(source);
+    if (scanner.errors.isNotEmpty) {
+      throw CumulatedTokenizerException(scanner.errors);
+    }
 
-    return tokens;
+    return scanner.tokens;
   }
 
-  Parser _createParser(
-    List<Token> tokens, {
-    AutoCompleteEngine? autoComplete,
+  (List<Token>, Parser, AutoCompleteEngine?) _createParser(
+    String source, {
+    bool autoComplete = false,
     bool? driftExtensions,
   }) {
-    final tokensForParser = tokens.where((t) => !t.invisibleToParser).toList();
-    return Parser(
+    final tokenizer = _scan(source);
+    final allTokens = tokenizer.tokens;
+    final tokensForParser =
+        allTokens.where((t) => !t.invisibleToParser).toList();
+    final autoCompleteEngine =
+        autoComplete ? AutoCompleteEngine(allTokens, this) : null;
+
+    final parser = Parser(
       tokensForParser,
       useDrift: driftExtensions ?? options.useDriftExtensions,
-      autoComplete: autoComplete,
+      autoComplete: autoCompleteEngine,
     );
+
+    for (final error in tokenizer.errors) {
+      parser.errors.add(ParsingError(error, error.message));
+    }
+
+    return (allTokens, parser, autoCompleteEngine);
   }
 
   /// Parses a single [sql] statement into an AST-representation.
   ParseResult parse(String sql) {
-    final tokens = tokenize(sql);
-    final parser = _createParser(tokens);
+    final (tokens, parser, _) = _createParser(sql);
 
     final stmt = parser.safeStatement();
     return ParseResult._(stmt, tokens, parser.errors, sql, null);
@@ -153,8 +171,7 @@ class SqlEngine {
   /// You can use the [AstNode.childNodes] of the returned [ParseResult.rootNode]
   /// to inspect the returned statements.
   ParseResult parseMultiple(String sql) {
-    final tokens = tokenize(sql);
-    final parser = _createParser(tokens);
+    final (tokens, parser, _) = _createParser(sql);
 
     final ast = parser.safeStatements();
     return ParseResult._(ast, tokens, parser.errors, sql, null);
@@ -165,8 +182,7 @@ class SqlEngine {
   /// The [ParseResult.rootNode] will be a [ColumnDefinition] with the parsed
   /// constraints.
   ParseResult parseColumnConstraints(String sql) {
-    final tokens = tokenize(sql);
-    final parser = _createParser(tokens, driftExtensions: false);
+    final (tokens, parser, _) = _createParser(sql, driftExtensions: false);
 
     return ParseResult._(
       ColumnDefinition(
@@ -186,8 +202,7 @@ class SqlEngine {
   /// The [ParseResult.rootNode] will either be a [TableConstraint] or an
   /// [InvalidStatement] in case of parsing errors.
   ParseResult parseTableConstraint(String sql) {
-    final tokens = tokenize(sql);
-    final parser = _createParser(tokens, driftExtensions: false);
+    final (tokens, parser, _) = _createParser(sql, driftExtensions: false);
 
     AstNode? constraint;
     try {
@@ -209,10 +224,8 @@ class SqlEngine {
   /// additional components like import statements.
   ParseResult parseDriftFile(String content) {
     assert(options.useDriftExtensions);
-
-    final tokens = tokenize(content);
-    final autoComplete = AutoCompleteEngine(tokens, this);
-    final parser = _createParser(tokens, autoComplete: autoComplete);
+    final (tokens, parser, autoComplete) =
+        _createParser(content, autoComplete: true);
 
     final driftFile = parser.driftFile();
     driftFile.scope = _constructRootScope();