+++ sqlparser/test/scanner/single_token_test.dart
@@ -1,3 +1,4 @@
+import 'package:sqlparser/sqlparser.dart';
 import 'package:sqlparser/src/reader/tokenizer/scanner.dart';
 import 'package:test/test.dart';
 
@@ -75,9 +76,7 @@ void main() {
   });
 
   test('issues error for unterminated string literals', () {
-    final scanner = Scanner("'unterminated");
-
-    expect(scanner.scanTokens, throwsA(isA<CumulatedTokenizerException>()));
+    final scanner = Scanner("'unterminated")..scanTokens();
 
     expect(
       scanner.errors,