+++ sqlparser/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.39.0-dev
+
+- When using the `parse` methods on `SqlEngine`, tokenizer errors are now
+  included in the result instead of throwing an exception.
+
 ## 0.38.1
 
 - Fix upcoming `unreachable_switch_default` lint.